Check Functions

Check functions for objects of class numeric, logical statements, class matching, value matching, file names, names, and named lists

check.numeric(x)

check.logic(
  x,
  statement,
  filternm = NULL,
  stopifnull = FALSE,
  stopifinvalid = TRUE,
  removeinvalid = FALSE,
  syntax = "R",
  returnvar = FALSE,
  returnpart = FALSE
)

check.matchclass(
  tab1,
  tab2,
  matchcol,
  var2 = NULL,
  tab1txt = NULL,
  tab2txt = NULL
)

check.matchval(
  tab1,
  tab2,
  var1,
  var2 = NULL,
  tab1txt = NULL,
  tab2txt = NULL,
  gui = FALSE,
  stopifmiss = FALSE,
  subsetrows = FALSE
)

checkfilenm(fn, outfolder = NULL, ext = NULL, stopifnull = FALSE)

checknm(nm, nmlst, ignore.case = TRUE)

check.namedlist(xlst, checknms = NULL, modetype = "numeric")

Value

The object who was checked, with necessary changes made for the object to pass the check (i.e. character vector to numeric vector).

Details

These functions are used to check if a vector is numeric with check.numeric, to check a logical statement with check.logic, to check if two objects' classes match, to check it two object's values match, to check file names, to check names, and finally, to check named lists.

Author

Tracey S. Frescino