List of population tables.

Returns a list of user-supplied parameters and parameter values for data evaluation (FIA or custom) extraction to be supplied to *DB functions.

eval_options(
  Cur = FALSE,
  Endyr = NULL,
  Endyr.filter = NULL,
  All = FALSE,
  Type = "VOL",
  evalid = NULL,
  invyrs = NULL,
  measyrs = NULL,
  varCur = "INVYR",
  evalType = NULL,
  ...
)

Arguments

Cur

Logical. If eval='FIA': extract plots with most current evaluation. If eval='custom': extract the most current sampled plots in the database.

Endyr

Integer (YYYY). If eval='FIA', defines end year for extracting one or more FIA evaluation. If eval='custom', defines end year for extracting the most current sampled plots until.

Endyr.filter

Filter. If endyr != NULL, a filter to identify when to use measEndyr, such as areas or plots identified as being disturbed in a particular year. In this example, plots sampled after the disturbance will be excluded.

All

Logical. If eval='FIA': includes all evaluations in database (annual inventory only). If eval='custom': includes all years in database (annual inventory only).

Type

String vector. Evaluation types ('ALL','CURR','VOL','P2VEG', DWM','INV','CHNG','GRM','REGEN'). If eval='FIA', Type is equivalent to plots for FIA Evaluations where 'ALL' includes nonsampled plots; 'CURR' and 'VOL' include plots used for area or tree estimates, respectively; Type = 'GRM' includes plots used for growth, removals, mortality; and Type = 'CHNG' includes plots used for change estimates (See FIA database manual for regioin availability and/or differences (https://www.fia.fs.usda.gov/library/database-documentation/index.php) If eval='custom', the associated tables are extracted for each Type. Multiple Types are accepted.

evalid

Integer. Only eval='FIA': extract data for a specific evaluation period. See notes for more information about FIA Evaluations.

invyrs

Integer vector. eval='custom': defines specific inventory years of data (e.g., 2010:2015). See FIA manual for definition of INVYR.

measyrs

Integer vector. eval='custom': defines specific measurement years of data (e.g., 2010:2015).

varCur

String. Name of variable to use for most current plot ('MEASYEAR', 'INVYR').

evalType

Deprecated. Use Type instead.

...

For extendibility.

Value

A list of user-supplied parameters and parameter values for strata.

Details

If no parameters, an empty list is returned.

Author

Tracey S. Frescino

Examples

eval_options(invyrs = 2015:2018)
#> $Cur
#> [1] FALSE
#> 
#> $Endyr
#> NULL
#> 
#> $Endyr.filter
#> NULL
#> 
#> $All
#> [1] FALSE
#> 
#> $Type
#> [1] "VOL"
#> 
#> $evalid
#> NULL
#> 
#> $invyrs
#> [1] 2015 2016 2017 2018
#> 
#> $measyrs
#> NULL
#> 
#> $varCur
#> [1] "INVYR"
#> 
#> $evalType
#> NULL
#>