Title output options.

Returns a list of user-supplied parameters and parameter values for outputting title with custom aesthetics.

title_options(
  title.main = NULL,
  title.ref = NULL,
  title.rowvar = NULL,
  title.colvar = NULL,
  title.unitvar = NULL,
  title.estvar = NULL,
  title.estvarn = NULL,
  title.filter = NULL,
  title.units = "acres",
  ...
)

Arguments

title.main

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: the complete title used for table. If title.main=NULL, the title.* parameters are used to generate title string. Note: if title.ref is not NULL, it is added to title.main.

title.ref

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: the ending text of the table title (e.g., Nevada, 2004-2005). If NULL, = "".

title.rowvar

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: pretty name for the row domain variable. If NULL, = rowvar.

title.colvar

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: pretty name for the column domain variable. If NULL, = colvar.

title.unitvar

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: pretty name for the estimation unit variable. If NULL, = unitvar.

title.estvar

String. TITLE: if savedata=TRUE and/or returntitle=TRUE: pretty name for the estimate variable. If NULL, title.estvar = estvar.name.

title.estvarn

String. TITLE: if savedata=TRUE and/or returntitle=TRUE: pretty name for the estimate variable. If NULL, title.estvar = estvar.name.

title.filter

String. TITLE, if savedata=TRUE and/or returntitle=TRUE: pretty name for filter(s). If title.filter=NULL, a default is generated from cfilter. If title.filter="", no title.filter is used.

title.units

String.

...

For extendibility.

Value

A list of user-supplied parameters and parameter values for outputting titles with custom aesthetics.

Details

If no parameters, an empty list is returned.

Author

Grayson W. White

Examples

title_options(title.main = "My fancy title", title.estvar = "Estimate title")
#> $title.main
#> [1] "My fancy title"
#> 
#> $title.ref
#> NULL
#> 
#> $title.rowvar
#> NULL
#> 
#> $title.colvar
#> NULL
#> 
#> $title.unitvar
#> NULL
#> 
#> $title.estvar
#> [1] "Estimate title"
#> 
#> $title.estvarn
#> NULL
#> 
#> $title.filter
#> NULL
#> 
#> $title.units
#> [1] "acres"
#>